Refactor units (#8)
* Move all units to a dedicated module

Add hierarchical CMakeLists

* Rename enums to CamelCase notation

* Basic Unit refactoring

- Add a new module "unit"
- Remove redundant Movable interface
- Extract attack logic to a separate module that can be injected
  (Strategy pattern)
- Extract Result enum to Attack module (maybe pull it out to a separate
  file?)

TODO:
- remove get_type from Unit

* Refactor Unit: remove dynamic polymorphism

- Replace Unit inheritance with std::variant
- Replace virtual functions with std::visit
- Store Units as std::optional<Unit> instead of a std::shared_ptr<Unit>
- Remove usage of get_type() throughout the code base

* Update unit counting logic

Move unit counting into a UnitCounter, where each class counts itself

* Format code using clang-format

Add clang-format to requirements.txt

* Remove max count unit map in Board

It is no longer needed after UnitCounter improvement

* Fix compilation errors

* Fix code formatting

* Add formatting script and hook

* Fix compilation issues #2

Add new script to apply_formatting

+++ include/ButtonInterface.hpp
@@ -4,31 +4,34 @@
 
 class ButtonInterface {
 protected:
-    int text_x_pos;
-    int text_y_pos;
-    bool highlighted;
-    const std::string path_to_textures;
-    sf::Sprite button;
-    sf::Sprite highlighted_button;
-    sf::Font font;
-    sf::Text button_text;
+  int text_x_pos;
+  int text_y_pos;
+  bool highlighted;
+  const std::string path_to_textures;
+  sf::Sprite button;
+  sf::Sprite highlighted_button;
+  sf::Font font;
+  sf::Text button_text;
 
-    void load_font();
-    void set_default_button_text(const std::string& text);
-    void update_text_pos();
+  void load_font();
+  void set_default_button_text(const std::string &text);
+  void update_text_pos();
 
 public:
-    explicit ButtonInterface(const std::string& text);
-    void set_text(const std::string& text);
-    void set_position(int x, int y);
-    void set_position(sf::Vector2f coords);
-    virtual void draw(sf::RenderWindow& win) = 0;
-    float get_height() const { return button.getGlobalBounds().height; };
-    float get_width() const { return button.getGlobalBounds().width; };
-    bool contains(int x, int y) const { return button.getGlobalBounds().contains(x, y); };
-    void highlight_on() { highlighted = true; };
-    void highlight_off() { highlighted = false; };
-    bool is_highlighted() const { return highlighted; };
-    sf::Vector2f get_position() const { return button.getPosition(); };
-    sf::String get_text() const { return button_text.getString(); };
+  explicit ButtonInterface(const std::string &text);
+  void set_text(const std::string &text);
+  void set_position(int x, int y);
+  void set_position(sf::Vector2f coords);
+  virtual void draw(sf::RenderWindow &win) = 0;
+  float get_height() const { return button.getGlobalBounds().height; };
+  float get_width() const { return button.getGlobalBounds().width; };
+  bool contains(int x, int y) const {
+    return button.getGlobalBounds().contains(static_cast<float>(x),
+                                             static_cast<float>(y));
+  };
+  void highlight_on() { highlighted = true; };
+  void highlight_off() { highlighted = false; };
+  bool is_highlighted() const { return highlighted; };
+  sf::Vector2f get_position() const { return button.getPosition(); };
+  sf::String get_text() const { return button_text.getString(); };
 };
